Split parsers into separate file
