+++ interface/mthalassa/include/mthalassa/mthalassa.h
@@ -1,0 +1,6 @@
+#ifndef MTHALASSA_MTHALASSA_H_
+#define MTHALASSA_MTHALASSA_H_
+
+#include <mthalassa/parsers.h>
+
+#endif
